Tear down the legacy spicy_regs.pipeline package
Now that the scheduled ETL runs entirely through the new
Pipeline/Reader/Writer path, retire the old Prefect pipeline and the
shared helpers it housed, so nothing imports spicy_regs.pipeline anymore.

Relocate the genuinely-shared code into the new structure:
- R2 connector: download_r2 + upload_r2 + load.py's R2 uploads folded
  into sources/r2.py (single self-contained connector; upload_file /
  upload_dataset / upload_comment_partitions / download_from_r2).
- S3 source helpers (get_agencies, list_json_files, download_and_parse)
  moved into sources/mirrulations.py, dropping the Prefect @task wrappers.
- BloomFilter + save_manifest moved into manifest.py (their only caller).
- The out-of-band enrich_pdf / backfill_derived_text CLIs moved to the
  top-level package alongside cli.py / mcp_server.py.

Then delete the package: the Prefect flow (pipeline.py), the back-compat
shims (transform.py, build_search_index.py), the relocated helpers, the
package README, and the legacy "Daily ETL Refresh" workflow + its `etl`
console script. Prefect is no longer a dependency.

Repoint all importers (transforms, schemas, the migrate script) and
tests; rewrite the manifest tests to exercise Manifest.load instead of
the removed PROCESSED_KEYS global; point deploy-docs' workflow_run at the
current "ETL (new pipeline)" workflow name.
