--- .github/workflows/etl.yml
@@ -1,86 +1,0 @@
-name: Daily ETL Refresh
-
-# The daily schedule has moved to the new Pipeline path ("ETL (new pipeline)").
-# This legacy workflow is kept as a manual fallback only — it is no longer
-# scheduled so it can never collide with the new pipeline's R2 writes.
-on:
-  workflow_dispatch:
-    inputs:
-      full_refresh:
-        description: 'Run full refresh instead of incremental'
-        required: false
-        default: false
-        type: boolean
-      batch_number:
-        description: 'Batch number (0-14) or leave empty for all batches'
-        required: false
-        default: ''
-        type: string
-
-jobs:
-  etl:
-    runs-on: ubuntu-latest
-    timeout-minutes: 60
-
-    steps:
-      - name: Determine batch number
-        id: batch
-        run: |
-          if [ "${{ github.event_name }}" = "schedule" ]; then
-            HOUR=$(date -u +%H)
-            BATCH=$((10#$HOUR - 6))
-            echo "number=$BATCH" >> $GITHUB_OUTPUT
-            echo "Running scheduled batch $BATCH (hour $HOUR UTC)"
-          elif [ -n "${{ inputs.batch_number }}" ]; then
-            echo "number=${{ inputs.batch_number }}" >> $GITHUB_OUTPUT
-            echo "Running manual batch ${{ inputs.batch_number }}"
-          else
-            echo "number=" >> $GITHUB_OUTPUT
-            echo "Running all agencies (no batch filter)"
-          fi
-
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.12'
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v4
-
-      - name: Install dependencies
-        run: uv sync
-
-      - name: Run incremental ETL
-        env:
-          R2_ACCESS_KEY_ID: ${{ secrets.R2_ACCESS_KEY_ID }}
-          R2_SECRET_ACCESS_KEY: ${{ secrets.R2_SECRET_ACCESS_KEY }}
-          R2_BUCKET_NAME: ${{ secrets.R2_BUCKET_NAME }}
-          R2_ENDPOINT: ${{ secrets.R2_ENDPOINT }}
-          R2_PUBLIC_URL: ${{ secrets.R2_PUBLIC_URL }}
-        run: |
-          BATCH_ARGS=""
-          POST_PROCESS=""
-          if [ -n "${{ steps.batch.outputs.number }}" ]; then
-            BATCH_ARGS="--batch-number ${{ steps.batch.outputs.number }} --batch-size 21"
-            # Only run feed summary + comment partitioning on the last batch
-            if [ "${{ steps.batch.outputs.number }}" != "14" ]; then
-              POST_PROCESS="--skip-post-process"
-            fi
-          fi
-
-          if [ "${{ inputs.full_refresh }}" = "true" ]; then
-            uv run etl --full-refresh $BATCH_ARGS $POST_PROCESS
-          else
-            uv run etl $BATCH_ARGS $POST_PROCESS
-          fi
-
-      - name: Upload artifacts (for debugging)
-        if: failure()
-        uses: actions/upload-artifact@v4
-        with:
-          name: etl-output-batch-${{ steps.batch.outputs.number }}
-          path: output/
-          retention-days: 7
