+++ pyproject.toml
@@ -12,7 +12,6 @@ dependencies = [
     "httpx",
     "duckdb>=1.4", # >=1.4 required to ATTACH an Iceberg REST catalog and MERGE INTO it
     "pandas",
-    "prefect",
     "cyclopts",
     "loguru",
     "mcp>=1.10.0",
@@ -44,10 +43,9 @@ docs = [
 ]
 
 [project.scripts]
-etl = "spicy_regs.pipeline.pipeline:app"
 run-pipeline = "spicy_regs.pipelines.regulations:app"
-enrich-pdf-text = "spicy_regs.pipeline.enrich_pdf:main"
-backfill-comment-text = "spicy_regs.pipeline.backfill_derived_text:main"
+enrich-pdf-text = "spicy_regs.enrich_pdf:main"
+backfill-comment-text = "spicy_regs.backfill_derived_text:main"
 embed = "spicy_regs.vectordb.embed:main"
 spicy-regs = "spicy_regs.cli:main"
 spicy-regs-mcp = "spicy_regs.mcp_server:main"
