+++ src/spicy_regs/enrich_pdf.py
