+++ src/spicy_regs/transforms/enrich_derived_text.py
@@ -11,7 +11,7 @@ anonymous S3 bucket the ETL already reads), needs no PDF download or parsing,
 and runs inline during staging — so ``comments.parquet`` ships with attachment
 text already populated. Comments whose attachment text is absent from
 derived-data are left untouched (``text_extraction_status`` stays ``None``) so
-the on-demand PDF-download fallback (:mod:`spicy_regs.pipeline.enrich_pdf`) can
+the on-demand PDF-download fallback (:mod:`spicy_regs.enrich_pdf`) can
 still backfill them.
 """
