+++ tests/test_bloom_filter.py
@@ -1,7 +1,7 @@
 """Tests for the BloomFilter used by the manifest loader."""
 
 
-from spicy_regs.pipeline.extract import BloomFilter
+from spicy_regs.manifest import BloomFilter
 
 
 class TestBloomFilter:
